feat: VLM support, multi-camera input, planner improvements
- core/LMP.py: VLM auto-detection (vl/vision/pixtral/llava/internvl), base64 JPEG image_url content blocks
- envs/robocasa_env.py: multi-camera RGB capture for VLM input (posed_person view)
- modules/interfaces.py: navigation/manipulation API extensions
- modules/planners.py: greedy planner stop_threshold/curvature handling
- prompts/robocasa_navigation/: avoidance with concrete examples, rotation/velocity refinements
- run_LMP.py: --vlm-cameras CLI, --obstacle-map-weight, --prompt-variant, parallel worker, eval metrics
- utils/utils.py: helper additions (image encoding, normalize)

+++ src/modules/planners.py
@@ -206,7 +206,7 @@ class PathPlanner:
         # Calculate all nearby voxel coordinates
         all_nearby_voxels = np.clip(current_pos + offsets_grid, 0, self.map_size - 1)
         # Remove duplicates, if any, caused by clipping
-        all_nearby_voxels = np.unique(all_nearby_voxels, axis=0)
+        all_nearby_voxels = np.unique(all_nearby_voxels, axis=0).astype(int)
         return all_nearby_voxels
     
     def _calculate_nearby_pixel(self, current_pos, object_centric=False):
@@ -220,7 +220,7 @@ class PathPlanner:
         # Calculate all nearby voxel coordinates
         all_nearby_voxels = np.clip(current_pos + offsets_grid, 0, self.map_size - 1)
         # Remove duplicates, if any, caused by clipping
-        all_nearby_voxels = np.unique(all_nearby_voxels, axis=0)
+        all_nearby_voxels = np.unique(all_nearby_voxels, axis=0).astype(int)
         return all_nearby_voxels
     
     def _postprocess_path(self, path, raw_target_map, object_centric=False):
